refactor(SelectCombo Component)
