+++ src/es6/constants.js
