+++ src/es6/contentIcons.js
