+++ src/es6/contexts.js
@@ -1,0 +1,7 @@
+import React from "react";
+import contextState from "../../state/state";
+
+export const AppState = React.createContext({
+  meta: contextState.meta || {},
+  project: null,
+});
