+++ src/es6/utils.js
